LayoutTests/imported/w3c/web-platform-tests/html/syntax/speculative-charset/speculative-script.tentative-expected.txt

CONSOLE MESSAGE: Error: assert_equals: Check result expected "token: 8cac0b7e-125c-4dd4-9e4d-7aeb8050e0a3, character: &#950;, previous character: ж, byte: ж" but got "token: 8cac0b7e-125c-4dd4-9e4d-7aeb8050e0a3, character: &#950;, previous character: NOT LOADED PREVIOUSLY, byte: ж"


FAIL Speculative script Error: assert_equals: Check result expected "token: 8cac0b7e-125c-4dd4-9e4d-7aeb8050e0a3, character: &#950;, previous character: ж, byte: ж" but got "token: 8cac0b7e-125c-4dd4-9e4d-7aeb8050e0a3, character: &#950;, previous character: NOT LOADED PREVIOUSLY, byte: ж"
